Fix Build issues for diagnostic tool!

+++ src/ControlStream.c
@@ -1657,7 +1657,7 @@ bool isControlDataInTransit(void) {
 bool LiShowMouseCursor(bool show){
     char *data = show ? "Relative" : "Absolute";
 
-    if(sendMessageAndForget(packetTypes[IDX_TOGGLE_MOUSE], strlen(data), data, CTRL_CHANNEL_UTF8, ENET_PACKET_FLAG_RELIABLE, false) == 0)
+    if(sendMessageAndForget(packetTypes[IDX_TOGGLE_MOUSE], (int)strlen(data), data, CTRL_CHANNEL_UTF8, ENET_PACKET_FLAG_RELIABLE, false) == 0)
     {
         Limelog("Error sending Mouse Mode on Control Stream.");
         return false;
